+++ src/main/java/com/gdsc/homework/controller/UserController.java
@@ -1,14 +1,12 @@
 package com.gdsc.homework.controller;
 
-import com.gdsc.homework.controller.dto.ResponseDTO;
+import com.gdsc.homework.controller.dto.response.ResponseDTO;
 import com.gdsc.homework.controller.dto.request.UserRequest;
-import com.gdsc.homework.controller.dto.response.UserDTO;
 import com.gdsc.homework.security.TokenProvider;
 import com.gdsc.homework.service.UserService;
 import com.gdsc.homework.service.dto.response.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -36,8 +34,7 @@ public class UserController {
         try {
             UserResponse userResponse = userService.signin(userRequest.toServiceDto());
             final String token = tokenProvider.createToken(userResponse);
-            UserDTO responseDTO = UserDTO.of(token, userResponse.getEmail(), userResponse.getNickName(), userResponse.getUserId());
-            return ResponseEntity.ok(responseDTO);
+            return ResponseEntity.ok(userResponse);
         } catch (Exception e) {
             ResponseDTO responseDTO = ResponseDTO.builder().error(e.getMessage()).build();
             return ResponseEntity.badRequest().body(responseDTO);
